style : 사이드바 및 전체 UI 개선
